@@ -1,6 +1,6 @@
 import {useCallback, useState}  from "react"
 import {api, keyHandler} from '../Utils'
-import {Redirect} from "react-router"
+import {Navigate} from "react-router"
 
 import {Button} from '../Buttons'
 import {TextInput, TextInputWithBtn} from '../Forms'
@@ -61,8 +61,8 @@ const LoginPage = (props) => {
     }, [email, password])
     
     // When the user already authenticated
-    if(isAuth()){
-        return <Redirect to={'/'}/>
+    if(props.isAuth){
+        return <Navigate to={'/'}/>
     }
     return (<>
         <div id="login-page" className="flex-col items-center content-center" style={backgroundStyle}>      
